migrate from Qt5 to Qt6
Linux developers will need to install the following packages:

apt install -y qt6-base-dev qt6-base-dev-tools libqt6opengl6-dev \
  libqt6websockets6-dev qtkeychain-qt6-dev

+++ src/proxy/AbstractTelnet.cpp
@@ -856,15 +856,15 @@ void AbstractTelnet::onReadInternal(const TelnetIacBytes &data)
     // so we parse the text and process all telnet commands:
 
     AppendBuffer cleanData;
-    cleanData.reserve(data.size());
+    cleanData.reserve(static_cast<int>(data.size()));
 
     int pos = 0;
     while (pos < data.size()) {
         if (m_inflateTelnet) {
             const int remaining = onReadInternalInflate(data.getQByteArray().data() + pos,
-                                                        data.size() - pos,
+                                                        static_cast<int>(data.size()) - pos,
                                                         cleanData);
-            pos = data.length() - remaining;
+            pos = static_cast<int>(data.length()) - remaining;
             // Continue because there might be additional chunks left to inflate
             continue;
         }
